- Http connection can be closed - StandardCharsets can be used
