+++ openwebstart/src/main/java/com/openwebstart/jvm/io/ConnectionUtils.java
@@ -7,6 +7,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
+import java.nio.charset.StandardCharsets;
 import java.security.DigestInputStream;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
@@ -35,12 +36,12 @@ public class ConnectionUtils {
     }
 
     public static String readUTF8Content(final InputStream inputStream) throws IOException {
-        return new String(IOUtils.readContent(inputStream), "UTF-8");
+        return new String(IOUtils.readContent(inputStream), StandardCharsets.UTF_8);
     }
 
     public static void writeUTF8Content(final OutputStream outputStream, final String content) throws IOException {
         Assert.requireNonNull(content, "content");
-        IOUtils.writeContent(outputStream, content.getBytes("UTF-8"));
+        IOUtils.writeContent(outputStream, content.getBytes(StandardCharsets.UTF_8));
     }
 
     public static InputStream getContentStream(final HttpURLConnection connection) throws IOException {
