+++ openwebstart/src/main/java/com/openwebstart/jvm/io/HttpResponse.java
@@ -3,40 +3,54 @@ package com.openwebstart.jvm.io;
 
 import net.adoptopenjdk.icedteaweb.Assert;
 
+import java.io.IOException;
 import java.io.InputStream;
+import java.net.HttpURLConnection;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class HttpResponse {
 
-    private final List<HttpHeader> headers;
+    private final HttpURLConnection connection;
 
-    private final int statusCode;
-
-    private final InputStream contentStream;
-
-    private final long contentSize;
-
-    public HttpResponse(final List<HttpHeader> headers, final int statusCode, final InputStream contentStream, final long contentSize) {
-        this.headers = Collections.unmodifiableList(Assert.requireNonNull(headers, "headers"));
-        this.statusCode = statusCode;
-        this.contentStream = Assert.requireNonNull(contentStream, "contentStream");
-        this.contentSize = contentSize;
+    public HttpResponse(final HttpURLConnection connection) {
+        this.connection = Assert.requireNonNull(connection, "connection");
     }
 
-    public List<HttpHeader> getHeaders() {
-        return headers;
+    private int readResponseCode() throws IOException {
+        return connection.getResponseCode();
     }
 
-    public InputStream getContentStream() {
-        return contentStream;
+    private String getResponseMessage() throws IOException {
+        return connection.getResponseMessage();
+    }
+
+    private HttpHeader getResponseHeader(final String name) {
+        return getResponseHeaders().stream()
+                .filter(h -> h.getName() != null)
+                .filter(h -> h.getName().equals(name))
+                .findFirst()
+                .orElse(null);
+    }
+
+    private List<HttpHeader> getResponseHeaders() {
+        return connection.getHeaderFields().
+                entrySet().
+                stream().
+                flatMap(e -> e.getValue().stream().map(v -> new HttpHeader(e.getKey(), v))).
+                collect(Collectors.toList());
     }
 
     public long getContentSize() {
-        return contentSize;
+        return connection.getContentLengthLong();
     }
 
-    public int getStatusCode() {
-        return statusCode;
+    public InputStream getContentStream() throws IOException {
+        return ConnectionUtils.getContentStream(connection);
+    }
+
+    public void closeConnection() {
+        connection.disconnect();
     }
 }
